Area of circle

+++ AreaOfCircle.java
@@ -3,7 +3,7 @@ public class AreaOfCircle{
 public static void main(String []args){
 Scanner input=new Scanner(System.in);
 System.out.println("Enter the radius :");
-double r=input.nextInt();
+double r=input.nextDouble();
 System.out.print("Area of the circle is :"+Math.PI*r*r);
 }
 }
